stubbing out more of the files

+++ DataStructures/HW01/Bank.h
@@ -30,26 +30,26 @@ namespace net_oatsnet_class_datastructures {
       ~Bank();
 
       /* Pre-condition: None
-       * Post-condition: Account is created if no errors exist, and BankError.NO_ERROR is returned
+       * Post-condition: Account is created if no errors exist, and NO_ERROR is returned
        * Params:
        *    string        name     : the account the name is filed under
        *    unsigned long checkings: the initial amount in the checkings account
        *    unsigned long savings  : the initial amount in the savings account
        * Returned error-conditions:
-       *    BankError.TOO_MANY_ACCOUNTS: There is not enough room to add another account (exceeds
+       *    TOO_MANY_ACCOUNTS: There is not enough room to add another account (exceeds
        *                                 MAX_ACCOUNTS)
-       *    BankError.DUPLICATE_NAME   : a duplicate name already exists in a previous account
-       *    BankError.EMPTY_STRING     : cannot add a name that is empty (e.g. "")
+       *    DUPLICATE_NAME   : a duplicate name already exists in a previous account
+       *    EMPTY_STRING     : cannot add a name that is empty (e.g. "")
        */
       int addAccount(string name, unsigned long checkings, unsigned long savings);
 
       /* Pre-condition: None
-       * Post-condition: Account with matching name is deleted, returns BankError.NO_ERROR
+       * Post-condition: Account with matching name is deleted, returns NO_ERROR
        * Params:
        *    string        name     : the account the name is filed under
        * Returned error-conditions:
-       *    BankError.ACCOUNT_NOT_FOUND: Given name is not in list of accounts
-       *    BankError.EMPTY_STRING     : cannot search for an empty name
+       *    ACCOUNT_NOT_FOUND: Given name is not in list of accounts
+       *    EMPTY_STRING     : cannot search for an empty name
        */
       int deleteAccount(string name);
 
@@ -59,21 +59,21 @@ namespace net_oatsnet_class_datastructures {
        *    The total value of all account types (checkings + savings) for every
        *    account that is held at the Bank is returned
        */
-      int getTotalAccountValues();
+      unsigned long long getTotalAccountValues();
 
       /* Pre-condition:
        *    accountType is using AccountType enum, and does not deviate from it
        * Post-condition: 
-       *    Account is credited with the correct amount if no errors, and BankError.NO_ERROR is returned
+       *    Account is credited with the correct amount if no errors, and NO_ERROR is returned
        *    If an error exists, nothing is to change
        * Params:
        *    string        name       : the account the name is filed under
        *    int           accountType: AccountType.CHECKINGS, AccountType.SAVINGS
        *    unsigned long amount     : amount the account is to be credited
        * Returned error-conditions:
-       *    BankError.ACCOUNT_NOT_FOUND: Given name is not in list of accounts
-       *    BankError.EMPTY_STRING     : cannot search for an empty name
-       *    BankError.OVERfLOW         : requested operation would result in an overflow
+       *    ACCOUNT_NOT_FOUND: Given name is not in list of accounts
+       *    EMPTY_STRING     : cannot search for an empty name
+       *    OVERfLOW         : requested operation would result in an overflow
        */
       int accountDeposit(string name, int accountType, unsigned long amount);
 
@@ -87,9 +87,9 @@ namespace net_oatsnet_class_datastructures {
        *    int           accountType: AccountType.CHECKINGS, AccountType.SAVINGS
        *    unsigned long amount     : amount the account is to be credited
        * Returned error-conditions:
-       *    BankError.ACCOUNT_NOT_FOUND : Given name is not in list of accounts
-       *    BankError.EMPTY_STRING      : cannot search for an empty name
-       *    BankError.INSUFFICIENT_FUNDS: requested accounts does not have enough funds
+       *    ACCOUNT_NOT_FOUND : Given name is not in list of accounts
+       *    EMPTY_STRING      : cannot search for an empty name
+       *    INSUFFICIENT_FUNDS: requested accounts does not have enough funds
        */
       int accountWithdraw(string name, int accountType, unsigned long amount);
